Bug 895526 - Part 0: refactor TestUpgradeRequired.
